fix(client): defer workspace metadata initialization
