+++ chat2db-community-client/src/layouts/init/init.tsx
@@ -4,18 +4,8 @@ import useDocumentListener from '@/hooks/useDocumentListener';
 import useCopyFocusData from '@/hooks/useFocusData';
 import useJavaMessageReceiver from '@/jcef/useProcessJavaPush';
 import miscServices from '@/service/misc';
-import supportedDatabaseService from '@/service/supportedDatabase';
 import { useGlobalStore } from '@/store/global';
 import { clearOlderLocalStorage } from '@/utils';
-import { buildIconSprite, registerDynamicDatabases } from '@/utils/dynamicDatabaseRegistry';
-import { databaseMap, databaseTypeList } from '@/constants/database';
-import {
-  dataSourceFormConfigs,
-  envItem,
-  portItem,
-  sshConfig,
-  storageItem,
-} from '@/components/ConnectionEdit/config/dataSource';
 import { isDesktop } from '@/utils/env';
 import clientRuntime from '@client-runtime';
 import { initializeDevEnvironmentIcon } from '@/utils/initLocalIcon';
@@ -99,28 +89,6 @@ const useInit = () => {
     registerMessage();
     registerNotification();
     initializeMonacoEditor();
-    // Surface backend configuration-only databases without a client rebuild.
-    supportedDatabaseService
-      .listSupported({})
-      .then((summaries) => {
-        const added = registerDynamicDatabases(
-          summaries,
-          { databaseMap, databaseTypeList, dataSourceFormConfigs },
-          { envItem, storageItem, portItem, sshConfig },
-        );
-        if (added.length) {
-          const sprite = buildIconSprite((summaries || []).filter((s) => added.includes(s.dbType)));
-          if (sprite && !document.getElementById('c2d-dynamic-db-icons')) {
-            const host = document.createElement('div');
-            host.id = 'c2d-dynamic-db-icons';
-            host.innerHTML = sprite;
-            document.body.appendChild(host);
-          }
-        }
-      })
-      .catch(() => {
-        // Older backends without the endpoint keep the built-in list.
-      });
   }, [serviceStatus, reload, isDesktop]);
 };
